# docs/contracts/http/commands.openapi.yaml
openapi: 3.1.0
info:
  title: HomeTusk Commands API
  version: 1.0.0
  description: |
    Intent-driven API for household task management.
    Users submit commands, not CRUD operations.

    **Stage 1:** Structured commands (no NL processing yet).

    See also:
    - [CLAUDE.md](../../../CLAUDE.md) — Development rules
    - [ADR-003](../../architecture/decisions/003-stage1-commands-api.md) — Stage 1 decisions
servers:
  - url: /api/v1
    description: API version 1
security:
  - bearerAuth: []
...
paths:
...
  /users/me:
    get:
      operationId: getCurrentUser
      summary: Get current user profile
      description: |
        Returns the current user's profile.
        If the user does not exist, creates a new profile from JWT claims.
      tags:
        - Users
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
  /households/{householdId}/tasks:
    get:
      operationId: listTasks
      summary: List tasks in a household
      tags:
        - Households
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, in_progress, done, cancelled]
        - name: assigneeId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: zoneId
          in: query
          required: false
          description: Filter by zone
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
  /households/{householdId}/tasks/{taskId}:
    get:
      operationId: getTask
      summary: Get a specific task
      tags:
        - Households
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetail'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
        '404':
          description: Task not found
...
  /households/{householdId}/notifications:
    get:
      operationId: listNotifications
      summary: List notifications for a household
      description: Returns notifications for the current user within the household.
      tags:
        - Notifications
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: since
          in: query
          required: false
          description: RFC3339 timestamp to filter notifications since this time
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          required: false
          description: Max number of results (default 50, max 200)
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Notification'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '403':
          description: Not a member of this household
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      summary: Mark notification as read
      tags:
        - Notifications
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Notification updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: Notification not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak JWT token
  schemas:
...
    CommandErrorResponse:
      type: object
      required:
        - correlationId
        - errorCode
        - message
      properties:
        correlationId:
          type: string
          format: uuid
        errorCode:
          type: string
          description: Machine-readable error code (see ErrorCode enum)
        message:
          type: string
          description: Human-readable error message
        validationErrors:
          type: array
          description: Present when errorCode is SCHEMA_INVALID
          items:
            $ref: '#/components/schemas/ValidationError'
        violations:
          type: array
          description: Present when errorCode is BUSINESS_RULE_VIOLATION
          items:
            $ref: '#/components/schemas/BusinessViolation'
    ValidationError:
      type: object
      required:
        - path
        - code
        - message
      properties:
        path:
          type: string
          description: JSON path to invalid field (e.g., $.payload.deadline)
        code:
          type: string
          description: Validation error code
        message:
          type: string
          description: Human-readable validation message
    BusinessViolation:
      type: object
      required:
        - rule
        - message
      properties:
        rule:
          type: string
          enum:
            - ASSIGNEE_MUST_BE_MEMBER
            - ZONE_MUST_EXIST
            - DEADLINE_MUST_BE_FUTURE
            - INITIATOR_MUST_HAVE_PERMISSION
            - TASK_ALREADY_COMPLETED
          description: Business rule that was violated
        message:
          type: string
          description: Human-readable explanation
    AuthErrorResponse:
      type: object
      required:
        - errorCode
        - message
      properties:
        errorCode:
          type: string
          enum:
            - AUTH_TOKEN_MISSING
            - AUTH_TOKEN_INVALID
            - AUTH_TOKEN_EXPIRED
        message:
          type: string

    # Domain entities
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        externalId:
          type: string
          description: Keycloak subject claim
        email:
          type: string
          format: email
        displayName:
          type: string
        avatarUrl:
          type: string
          format: uri
        households:
          type: array
          items:
            $ref: '#/components/schemas/HouseholdSummary'
        createdAt:
          type: string
          format: date-time
    HouseholdSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        role:
          type: string
          enum: [admin, member]
...
    Zone:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        householdId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, done, cancelled]
        assignee:
          $ref: '#/components/schemas/UserSummary'
        zone:
          $ref: '#/components/schemas/Zone'
        deadline:
          type: string
          format: date-time
        createdBy:
          $ref: '#/components/schemas/UserSummary'
        commandId:
          type: string
          format: uuid
          description: Command that created this task
        createdVia:
          type: string
          enum: [command, fallback, direct]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string

    # Request schemas
...
    Notification:
      type: object
      required:
        - id
        - householdId
        - userId
        - type
        - payload
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        householdId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
          enum:
            - invite_accepted
            - task_assigned
            - task_completed
            - shopping_item_added
            - shopping_item_purchased
        payload:
          $ref: '#/components/schemas/NotificationPayload'
        createdAt:
          type: string
          format: date-time
        readAt:
          type: string
          format: date-time
          nullable: true
        correlationId:
          type: string
          format: uuid
          nullable: true
    NotificationPayload:
      type: object
      properties:
        actorUserId:
          type: string
          format: uuid
        entityId:
          type: string
          format: uuid
        entityType:
          type: string
          description: Entity type (task, shopping_item, invite)
        summary:
          type: string
          description: Human-readable summary

    # Shopping schemas
...
    ShoppingItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        quantity:
          type: integer
        unit:
          type: string
        purchased:
          type: boolean
        linkedTaskId:
          type: string
          format: uuid
          description: Associated task (if any)
        addedBy:
          $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
        purchasedAt:
          type: string
          format: date-time
...
    TaskDetail:
      allOf:
        - $ref: '#/components/schemas/Task'
        - type: object
          properties:
            linkedShoppingItems:
              type: array
              items:
                $ref: '#/components/schemas/ShoppingItem'